ingestor/requirements.txt
paho-mqtt==2.1.0
fastapi[standard]==0.115.6
sqlalchemy==2.0.36
pydantic-settings==2.7.1
pydantic==2.10.6
aiofiles==24.1.0
alembic==1.15.1
psycopg2-binary==2.9.10
requests==2.32.5
fermento-embedded-schemas @ git+https://github.com/t3knomanzer/fermento-services.git@master#subdirectory=packages/fermento-embedded-schemas
fermento-service-schemas @ git+https://github.com/t3knomanzer/fermento-services.git@master#subdirectory=packages/fermento-service-schemas